Efforts were made to slim the main py file new Alignement.py created with     alignSingle()     ScoreSingle()     getSequenceCentroid()     alignSequences()     slidingWindow()     dictToFile() new params.py created with     the creation of the yaml variables     accessible by:         import Params as p         p.<some_parameter>
